Added prel_pop() and active/inactive status on queue elements
The idea is to be able to fetch an item from a queue, without removing it,
but only marking it as 'inactive'. The functions list_queue() and pop() will
automatically skip inactive records (corresponding functions that also list
inactive records are needed, but haven't yet been written).

prel_pop(Db, Table, Q) -> {ok, Object, ActualKey} | done

The ActualKey can be passed to extract/4 or delete/3 in order to permanently
remove the object.