added dump_tables/1; indexes; prel_pop now sets obj status to 'blocking'
kvdb:dump_tables(Name) is mainly a debugging function for use on small
test databases.

Attributes can now be indexed. See documentation (kvdb.md)

When popping something from a queue using prel_pop, the status of the object
is no longer set to 'disabled', but to 'blocking'. This means that any
subsequent pop operations before the object has been deleted, will result
in the return value 'blocked'. The purpose of this is to ensure that commands
are executed in the order in which they are queued. A blocking entry must
either be retried until successful and then deleted, or marked as 'disabled',
in which case it will be skipped (this last feature is basically implemented
but not yet exported or fully tested).

Documentation has been improved somewhat.

Apologies for the massive commit.