Indexing now possible on value part, as well as via callback.
If, instead of naming an attribute, one specifies {value},
the value part of the object is used for the indexing operation.
It is only fetched (once) if needed - that is, if any of the
indexing operations reference the value part.

If instead, {M,F} is specified, then M:F({Key, Attrs, Value})
becomes the basis for the indexing operation.